ADDING SECURITY FEATURES

+++ src/main/java/com/example/movieApi/dto/MoviePageResponse.java
@@ -5,8 +5,8 @@ import java.util.List;
 public record MoviePageResponse(List<MovieDto> movieDto,
                                Integer pageNumber,
                                Integer pageSize,
-                               int totalElements,
-                               int totalPages,
+                               long totalElements,
+                               long totalPages,
                                Boolean isLast) {
 
 }
